BINIUS-120: split M4 into binius-m4-verifier and binius-m4-prover
Address review: model the M4 code on the Prover/Verifier split in
binius-prover / binius-verifier, where the verifier crate is the base
and the prover depends on it.

Structure now mirrors the convention exactly:
- binius-m4-verifier (base): BatchCommitLayout (shared shape) + Verifier.
- binius-m4-prover (depends on it): ValueTable (witness) + Prover.

The witness moves to the prover crate, since the verifier never touches
a witness and ValueTable commits through an IOPProverChannel.

Per the review comments:
- Drop oracle_spec() from the layout: the layout knows only shape, not
  the protocol. The oracle spec is built in Verifier::setup, and it is
  non-ZK (OracleSpec::new), since ZK-ness is a higher-level choice.
- Lift one-time work into setup. Verifier::setup builds the Merkle
  scheme, FRI parameters, and verifier compiler once. Prover::setup
  inherits them via from_verifier_compiler and uses the multithreaded
  NTT, not the single-threaded one.
