+++ crates/m4-prover/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "binius-m4-prover"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+binius-core = { path = "../core" }
+binius-field = { path = "../field" }
+binius-frontend = { path = "../frontend" }
+binius-hash = { path = "../hash" }
+binius-iop-prover = { path = "../iop-prover" }
+binius-ip-prover = { path = "../ip-prover" }
+binius-m4-verifier = { path = "../m4-verifier" }
+binius-math = { path = "../math" }
+binius-transcript = { path = "../transcript" }
+binius-utils = { path = "../utils" }
+binius-verifier = { path = "../verifier" }
+
+[dev-dependencies]
+assert_matches = { workspace = true }
+binius-iop = { path = "../iop" }
+binius-prover = { path = "../prover" }
+proptest = { workspace = true }
+rand = { workspace = true }
+
+[features]
+default = ["rayon"]
+rayon = ["binius-utils/rayon"]
