+++ crates/m4-prover/src/lib.rs
@@ -1,0 +1,9 @@
+// Copyright 2025 Irreducible Inc.
+
+//! Witness table and prover for the data-parallel Binius64 M4 proof system.
+
+mod prove;
+mod value_table;
+
+pub use prove::Prover;
+pub use value_table::{PopulateInstanceError, ValueTable};
