--- crates/m4/src/lib.rs
@@ -1,11 +1,0 @@
-// Copyright 2025 Irreducible Inc.
-
-//! Witness table and commitment for the data-parallel Binius64 M4 proof system.
-
-mod commit;
-mod pcs;
-mod value_table;
-
-pub use commit::BatchCommitLayout;
-pub use pcs::BatchPcsParams;
-pub use value_table::{PopulateInstanceError, ValueTable};
